# openapi.yaml
openapi: 3.1.0
info:
  title: MAX Business Bot API
  version: 0.1.0
paths:
  /api/me:
    get:
      tags:
      - miniapp
      summary: Текущий пользователь мини-приложения
      operationId: me_api_me_get
      parameters:
      - name: x-max-init-data
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Max-Init-Data
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/events:
    post:
      tags:
      - miniapp
      summary: Событие аналитики из мини-приложения
      operationId: post_event_api_events_post
      parameters:
      - name: x-max-init-data
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Max-Init-Data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackRequest'
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /health:
    get:
      tags:
      - service
      summary: Проверка, что сервис жив
      operationId: health_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Health Health Get
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    MeResponse:
      properties:
        user_id:
          type: integer
          title: User Id
        first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
        is_dev:
          type: boolean
          title: Is Dev
          default: false
      type: object
      required:
      - user_id
      title: MeResponse
    TrackRequest:
      properties:
        name:
          type: string
          title: Name
        props:
          additionalProperties: true
          type: object
          title: Props
          default: {}
      type: object
      required:
      - name
      title: TrackRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError